Move components/diff-viewer to version-control
