+++ src/components/layout/main-layout.tsx
@@ -9,6 +9,7 @@ import { usePersistentSettingsStore } from "@/settings/stores/persistent-setting
 import { useSettingsStore } from "@/settings/stores/settings-store";
 import { useBufferStore } from "@/stores/buffer-store";
 import { useUIState } from "@/stores/ui-state-store";
+import DiffViewer from "@/version-control/diff-viewer/views/diff-viewer";
 import { type GitHunk, stageHunk, unstageHunk } from "../../version-control/git/controllers/git";
 import AIChat from "../ai-chat/ai-chat";
 import BottomPane from "../bottom-pane";
@@ -16,7 +17,6 @@ import CommandBar from "../command/components/command-bar";
 import CommandPalette from "../command/components/command-palette";
 import ThemeSelector from "../command/components/theme-selector";
 import type { Diagnostic } from "../diagnostics/diagnostics-pane";
-import DiffViewer from "../diff-viewer/diff-viewer";
 import CodeEditor from "../editor/code-editor";
 import EditorFooter from "../editor-footer";
 import ExtensionsView from "../extensions-view";
