+++ src/version-control/diff-viewer/views/diff-hunk-header.tsx
@@ -1,7 +1,7 @@
 import { ChevronDown, ChevronRight, Copy, Minus, Plus } from "lucide-react";
 import { cn } from "@/utils/cn";
-import { copyLineContent, createGitHunk } from "./utils/diff-helpers";
-import type { DiffHunkHeaderProps } from "./utils/types";
+import { copyLineContent, createGitHunk } from "../controllers/diff-helpers";
+import type { DiffHunkHeaderProps } from "../models/diff-types";
 
 export function DiffHunkHeader({
   hunk,
