+++ src/version-control/diff-viewer/views/multi-file-diff-viewer.tsx
@@ -1,10 +1,10 @@
 import { ChevronDown, ChevronRight, FileIcon, FilePlus, FileText, FileX } from "lucide-react";
 import { memo, useCallback, useMemo, useState } from "react";
 import type { GitDiff } from "@/version-control/git/controllers/git";
+import type { FileDiffSummary, MultiFileDiffViewerProps } from "../models/diff-types";
 import { DiffHeader } from "./diff-header";
 import { ImageDiffViewer } from "./image-diff-viewer";
 import { TextDiffViewer } from "./text-diff-viewer";
-import type { FileDiffSummary, MultiFileDiffViewerProps } from "./utils/types";
 
 // Memoized file component to prevent unnecessary re-renders
 interface FileRowProps {
